Modify the Dockerfile to create and use a local "jekyll" user to prevent the rvm, gem installs at the root level and the jekyll site from being run at root.
Note that the rvm install does require sudo access to install
dependencies using apt, and for this, the Dockerfile also
includes installs sudo and adding a file under /etc/sudoer.d/
for this permission.

Modify also the README.md file accordingly, for the destination
mount to be the "site" directory under the "jekyll" user, and
also to expose, instead of publishing or binding port 4000
to the host. By default, with the publish parameter, docker
opens up access to the forwarded port to any IP address (0.0.0.0)
on the host. Using the expose parameter only, the port is not
opened on the host.